backend/requirements.txt
Django==4.2.28
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-extensions==3.2.3
pymongo==4.6.0
python-dotenv==1.0.0
pillow>=10.3.0
requests==2.31.0
gunicorn==21.2.0
